♻️: 라이브코딩 예외처리 수정

+++ src/main/java/com/sw/cmc/adapter/out/lcd/persistence/LiveCodingRepositoryImpl.java
@@ -1,6 +1,7 @@
 package com.sw.cmc.adapter.out.lcd.persistence;
 
 import com.sw.cmc.application.service.redis.RedisService;
+import com.sw.cmc.common.advice.CmcException;
 import com.sw.cmc.domain.lcd.LiveCodingDomain;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -38,15 +39,24 @@ public class LiveCodingRepositoryImpl implements LiveCodingRepository {
 
         // Redis에 방 정보 저장
         redisService.saveHash(REDIS_LIVE_CODING_PREFIX + liveCodingDomain.getRoomId().toString(), liveCodingMap);
+
+        //  hostId 기준으로 roomId 저장
+        redisService.save(REDIS_LIVE_CODING_PREFIX + "host:" + liveCodingDomain.getHostId(), liveCodingDomain.getRoomId().toString());
     }
 
     @Override
     public boolean deleteLiveCoding(UUID roomId) {
-        // Redis에서 방을 삭제하는 로직
-        String key = "live_coding:" + roomId;  // Redis에서 사용할 key
+        String key = REDIS_LIVE_CODING_PREFIX + roomId;  // 방 정보 저장 키
+        Map<String, String> liveCodingMap = redisService.selectHash(key);  // 방 정보 조회
+        if (liveCodingMap == null || liveCodingMap.isEmpty()) {
+            throw new CmcException("LCD001");
+        }
+
+        String hostIdKey = REDIS_LIVE_CODING_PREFIX + "host:" + liveCodingMap.get("hostId");
+        redisService.delete(hostIdKey);  // 호스트 ID 기반 매핑 삭제
+
         return redisService.delete(key);  // Redis에서 해당 key 삭제
     }
-
 
     @Override
     public LiveCodingDomain findByRoomId(UUID roomId) {
@@ -71,5 +81,17 @@ public class LiveCodingRepositoryImpl implements LiveCodingRepository {
 
         return new LiveCodingDomain(retrievedRoomId, hostId, createdAt, participantCount, participants);
     }
+
+    @Override
+    public UUID findRoomIdByHostId(Long hostId) {
+        String roomIdStr = redisService.select(REDIS_LIVE_CODING_PREFIX + "host:" + hostId);
+        return roomIdStr != null ? UUID.fromString(roomIdStr) : null;
+    }
+
+    @Override
+    public boolean existsByHostId(Long hostId) {
+        return findRoomIdByHostId(hostId) != null;
+    }
+
 
 }
